pyproject.toml
[project]
name = "beeai-agents"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "beeai-sdk>=0.3.2",
    "pyyaml>=6.0.2",
    "beeai-framework",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[project.scripts]
server = "beeai_agents.agent:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
